LPS-145913 Move resource actions method and update test
